fix(java): Line up protobuf type handling with our other generators

+++ generators/java/generator-utils/src/main/java/com/fern/java/PoetTypeNameMapper.java
@@ -161,17 +161,17 @@ public final class PoetTypeNameMapper {
         @Override
         public TypeName visitUint() {
             if (primitiveAllowed) {
-                return TypeName.LONG;
+                return TypeName.INT;
             }
-            return ClassName.get(Long.class);
+            return ClassName.get(Integer.class);
         }
 
         @Override
         public TypeName visitUint64() {
             if (primitiveAllowed) {
-                return TypeName.LONG;
+                return TypeName.INT;
             }
-            return ClassName.get(Long.class);
+            return ClassName.get(Integer.class);
         }
 
         @Override
